[v 1.0.0 ] [ 🍀 - PR Comment 반영: Platform 모듈 분리 및 Firebase SDK 의존성을 덜어내기 위한 래퍼 생성 ]
+) Platform 폴더 구조 변경
+) Error -> Result로 네이밍 변경
