+++ iTime/Platform/Sources/Platform/Endpoint/DatabaseEndpoint.swift
@@ -1,0 +1,20 @@
+//
+//  File.swift
+//  
+//
+//  Created by 이상헌 on 2023/10/27.
+//
+
+import Foundation
+
+// MARK: - DocumentReferenceConvertible
+
+public protocol DocumentReferenceConvertible {
+    var referencePath: String { get }
+}
+
+// MARK: - DatabaseEndpoint
+
+public enum DatabaseEndpoint: String {
+    case tests = "test_v2"
+}
